서비스 분리 및 Provider 구현
- ILocation, IObject 서비스 강등 및 Model에 병합 예정

+++ Assets/Scripts/GameSystem/SaveManager.cs
@@ -19,7 +19,7 @@ public class SaveManager : MonoBehaviour
 
     public void LoadDaySave(int day)
     {
-        dataService.LoadDay((uint)day);
+        dataService.LoadDay(day);
         SceneManager.LoadScene("DayLoading");
     }
 }
